refactor(observables): modularize Observable statics
