+++ src/observables/CallbackObservable.ts
@@ -108,3 +108,5 @@ function dispatchNext({ value, subscriber }) {
 function dispatchError({ err, subscriber }) {
   subscriber.error(err);
 }
+
+Observable.fromCallback = CallbackObservable.create;
