+++ src/observables/EmptyObservable.ts
@@ -26,3 +26,5 @@ export class EmptyObservable<T> extends Observable<T> {
     }
   }
 }
+
+Observable.empty = EmptyObservable.create;
