+++ src/observables/ErrorObservable.ts
@@ -29,3 +29,5 @@ export class ErrorObservable<T> extends Observable<T> {
     }
   }
 }
+
+Observable.throw = ErrorObservable.create;
