+++ src/observables/FromEventObservable.ts
@@ -51,3 +51,5 @@ export class FromEventObservable<T, R> extends Observable<T> {
     FromEventObservable.setupSubscription(sourceObj, eventName, handler, subscriber);
   }
 }
+
+Observable.fromEvent = FromEventObservable.create;
