+++ src/observables/FromObservable.ts
@@ -46,3 +46,5 @@ export class FromObservable<T> extends Observable<T> {
     }
   }
 }
+
+Observable.from = FromObservable.create;
