+++ src/observables/IntervalObservable.ts
@@ -43,3 +43,5 @@ export class IntervalObservable<T> extends Observable<T> {
     }));
   }
 }
+
+Observable.interval = IntervalObservable.create;
