+++ src/observables/PromiseObservable.ts
@@ -66,3 +66,5 @@ function dispatchNext({ value, subscriber }) {
 function dispatchError({ err, subscriber }) {
   subscriber.error(err);
 }
+
+Observable.fromPromise = PromiseObservable.create;
