+++ src/observables/TimerObservable.ts
@@ -67,3 +67,5 @@ export class TimerObservable<T> extends Observable<T> {
     subscriber.add(scheduler.schedule(TimerObservable.dispatch, dueTime, { index, period, subscriber }));
   }
 }
+
+Observable.timer = TimerObservable.create;
